+++ Src/FluentAssertions/Formatting/Formatter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions.Common;
 using FluentAssertions.Equivalency;
+using FluentAssertions.Execution;
 using FluentAssertions.Xml;
 
 namespace FluentAssertions.Formatting
@@ -68,14 +69,16 @@ namespace FluentAssertions.Formatting
         /// Returns a human-readable representation of a particular object.
         /// </summary>
         /// <param name="value">The value for which to create a <see cref="string"/>.</param>
-        /// <param name="useLineBreaks">
+        /// <param name="options">
         /// Indicates whether the formatter should use line breaks when the specific <see cref="IValueFormatter"/> supports it.
         /// </param>
         /// <returns>
         /// A <see cref="string" /> that represents this instance.
         /// </returns>
-        public static string ToString(object value, bool useLineBreaks = false)
+        public static string ToString(object value, FormattingOptions options = null)
         {
+            options ??= new FormattingOptions();
+
             try
             {
                 if (isReentry)
@@ -90,11 +93,21 @@ namespace FluentAssertions.Formatting
 
                 var context = new FormattingContext
                 {
-                    Depth = graph.Depth,
-                    UseLineBreaks = useLineBreaks
+                    UseLineBreaks = options.UseLineBreaks
                 };
 
-                return Format(value, context, (path, childValue) => FormatChild(path, childValue, useLineBreaks, graph));
+                FormattedObjectGraph output = new(options.MaxLines);
+
+                try
+                {
+                    Format(value, output, context, (path, childValue,
+                        output) => FormatChild(path, childValue, output, context, options, graph));
+                }
+                catch (MaxLinesExceededException)
+                {
+                }
+
+                return output.ToString();
             }
             finally
             {
@@ -102,29 +115,28 @@ namespace FluentAssertions.Formatting
             }
         }
 
-        private static string FormatChild(string path, object childValue, bool useLineBreaks, ObjectGraph graph)
+        private static void FormatChild(string path, object value, FormattedObjectGraph output, FormattingContext context, FormattingOptions options, ObjectGraph graph)
         {
             try
             {
                 Guard.ThrowIfArgumentIsNullOrEmpty(path, nameof(path), "Formatting a child value requires a path");
 
-                if (!graph.TryPush(path, childValue))
+                if (!graph.TryPush(path, value))
                 {
-                    return $"{{Cyclic reference to type {childValue.GetType()} detected}}";
+                    output.AddFragment($"{{Cyclic reference to type {value.GetType()} detected}}");
                 }
-                else if (graph.Depth > 5)
+                else if (graph.Depth > options.MaxDepth)
                 {
-                    return "{Maximum recursion depth was reached…}";
+                    output.AddLine($"Maximum recursion depth of {options.MaxDepth} was reached. " +
+                           $" Increase {nameof(FormattingOptions.MaxDepth)} on {nameof(AssertionScope)} or {nameof(AssertionOptions)} to get more details.");
                 }
                 else
                 {
-                    var context = new FormattingContext
+                    using (output.WithIndentation())
                     {
-                        Depth = graph.Depth,
-                        UseLineBreaks = useLineBreaks
-                    };
-
-                    return Format(childValue, context, (x, y) => FormatChild(x, y, useLineBreaks, graph));
+                        Format(value, output, context, (childPath, childValue, nestedOutput) =>
+                            FormatChild(childPath, childValue, nestedOutput, context, options, graph));
+                    }
                 }
             }
             finally
@@ -133,10 +145,10 @@ namespace FluentAssertions.Formatting
             }
         }
 
-        private static string Format(object value, FormattingContext context, FormatChild formatChild)
+        private static void Format(object value, FormattedObjectGraph output, FormattingContext context, FormatChild formatChild)
         {
             IValueFormatter firstFormatterThatCanHandleValue = Formatters.First(f => f.CanHandle(value));
-            return firstFormatterThatCanHandleValue.Format(value, context, formatChild);
+            firstFormatterThatCanHandleValue.Format(value, output, context, formatChild);
         }
 
         /// <summary>