+++ Src/FluentAssertions/Formatting/IValueFormatter.cs
@@ -1,4 +1,6 @@
-﻿namespace FluentAssertions.Formatting
+﻿using System.Collections.Generic;
+
+namespace FluentAssertions.Formatting
 {
     /// <summary>
     /// Represents a strategy for formatting an arbitrary value into a human-readable string representation.
@@ -9,7 +11,8 @@
     public interface IValueFormatter
     {
         /// <summary>
-        /// Indicates whether the current <see cref="IValueFormatter"/> can handle the specified <paramref name="value"/>.
+        /// Indicates
+        /// whether the current <see cref="IValueFormatter"/> can handle the specified <paramref name="value"/>.
         /// </summary>
         /// <param name="value">The value for which to create a <see cref="string"/>.</param>
         /// <returns>
@@ -20,38 +23,22 @@
         /// <summary>
         /// Returns a human-readable representation of <paramref name="value"/>.
         /// </summary>
-        /// <param name="value">The value for which to format.</param>
+        /// <param name="value">The value to format into a human-readable representation</param>
+        /// <param name="formattedGraph">
+        ///     An object to write the textual representation to.
+        /// </param>
         /// <param name="context">
-        /// Contains additional information about the formatting task.
+        ///     Contains additional information that the implementation should take into account.
         /// </param>
         /// <param name="formatChild">
-        /// Allows the formatter to recursively format any child objects.
+        ///     Allows the formatter to recursively format any child objects.
         /// </param>
         /// <remarks>
-        /// DO NOT CALL <see cref="Formatter.ToString(object,bool)"/> directly, but use <paramref name="formatChild"/>
+        /// DO NOT CALL <see cref="Formatter.ToString(object,FormattingOptions)"/> directly, but use <paramref name="formatChild"/>
         /// instead. This will ensure cyclic dependencies are properly detected.
+        /// Also, the <see cref="FormattedObjectGraph"/> may throw
+        /// an <see cref="MaxLinesExceededException"/> that must be ignored by implementations of this interface.
         /// </remarks>
-        string Format(object value, FormattingContext context, FormatChild formatChild);
+        void Format(object value, FormattedObjectGraph formattedGraph, FormattingContext context, FormatChild formatChild);
     }
-
-    /// <summary>
-    /// Provides information about the current formatting action.
-    /// </summary>
-    public class FormattingContext
-    {
-        public int Depth { get; set; }
-
-        public bool UseLineBreaks { get; set; }
-    }
-
-    /// <summary>
-    /// Represents a method that can be used to format child values from inside a <see cref="IValueFormatter"/>.
-    /// </summary>
-    /// <param name="childPath">
-    /// Represents the path from the current location to the child value.
-    /// </param>
-    /// <param name="value">
-    /// The child value to run through the configured <see cref="IValueFormatter"/>s.
-    /// </param>
-    public delegate string FormatChild(string childPath, object value);
 }