+++ Src/FluentAssertions/Formatting/StringValueFormatter.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace FluentAssertions.Formatting
+﻿namespace FluentAssertions.Formatting
 {
     public class StringValueFormatter : IValueFormatter
     {
@@ -16,13 +14,18 @@ namespace FluentAssertions.Formatting
             return value is string;
         }
 
-        /// <inheritdoc />
-        public string Format(object value, FormattingContext context, FormatChild formatChild)
+        public void Format(object value, FormattedObjectGraph formattedGraph, FormattingContext context, FormatChild formatChild)
         {
-            string prefix = context.UseLineBreaks ? Environment.NewLine : string.Empty;
-            string escapedString = value.ToString();
+            string result = "\"" + value + "\"";
 
-            return prefix + "\"" + escapedString + "\"";
+            if (context.UseLineBreaks)
+            {
+                formattedGraph.AddFragmentOnNewLine(result);
+            }
+            else
+            {
+                formattedGraph.AddFragment(result);
+            }
         }
     }
 }